backend/pyproject.toml
[tool.poetry]
name = "file-sharing-app"
version = "0.1.0"
description = ""
authors = ["vusallyv <vusal581176@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "4.0.5"
amqp = "^5.1.1"
celery = "^5.2.7"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
django-rest-swagger = "^2.2.0"
django-celery-beat = "2.3.0"
channels = "^3.0.5"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
